refactor: search styles
